requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
defusedxml==0.7.1
Django==3.2.9
djangorestframework==3.12.4
httpie==2.6.0
idna==3.3
Pygments==2.10.0
PySocks==1.7.1
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.2
typing-extensions==3.10.0.0
urllib3==1.26.7
